engine/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "agentport-engine"
version = "0.1.0"
description = "AgentPort code-generation engine"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = ["pytest"]

[project.scripts]
agentport = "agentport_engine._cli:main"

[tool.setuptools.packages.find]
where = ["."]
